backend/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-stubs==1.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==1.1.0
djoser==2.0.3
mypy==0.761
mypy-extensions==0.4.3
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
